% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.defol}
\alias{as.defol}
\title{Alias to [as_defol()]}
\usage{
as.defol(x)
}
\arguments{
\item{x}{A data frame or list-like object to cast. Must have named elements
for "year", "series", "gsi", "ngsi", and "defol_status".}
}
\value{
`x` cast to a `defol` object
}
\description{
Alias to [as_defol()]
}
\examples{
data(dmj_defol)
example_data <- as.data.frame(dmj_defol)
is.defol(example_data)
back_to_defol <- as_defol(example_data)
is.defol(back_to_defol)

}
