% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_defol}
\alias{as_defol}
\title{Cast data frame to list-like \code{defol} object}
\usage{
as_defol(x)
}
\arguments{
\item{x}{A data frame or list-like object to cast. Must have named elements
for "year", "series", "gsi", "ngsi", and "defol_status".}
}
\value{
\code{x} cast to a \code{defol} object
}
\description{
Cast data frame to list-like \code{defol} object
}
\examples{
data(dmj_defol)
example_data <- as.data.frame(dmj_defol)
is.defol(example_data)
back_to_defol <- as_defol(example_data)
is.defol(back_to_defol)

}
