% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_obr}
\alias{as_obr}
\title{Cast data frame to list-like \code{obr} object}
\usage{
as_obr(x)
}
\arguments{
\item{x}{A data frame or list-like object to cast. Must have named elements
for "year", "samp_depth", "num_defol", "perc_defol", "num_max_defol",
"perc_max_defol", "mean_gsi", "mean_ngsi", "outbreak_status".}
}
\value{
\code{x} cast to an \code{obr} object
}
\description{
Cast data frame to list-like \code{obr} object
}
\examples{
data(dmj_obr)
example_data <- as.data.frame(dmj_obr)
is.obr(example_data)
back_to_obr <- as_obr(example_data)
is.obr(back_to_obr)

}
