% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_host_series}
\alias{correct_host_series}
\title{Correct host series}
\usage{
correct_host_series(input_series)
}
\arguments{
\item{input_series}{data.frame rwl object with the host tree series
as first column and non-host chornology as second}
}
\value{
data.frame with 3 added columns, (1) the mean/sd sdjusted non-host chronology,
(2) the "corrected" host series after subtraction from the adjusted host chronology,
and (3) the "normalized" or scaled climate-corrected host series used to identify defoliation events
}
\description{
Removes the nonhost growth signal from host series
}
