% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defol}
\alias{defol}
\title{Constructor for S3 defol class}
\usage{
defol(year, series, gsi, ngsi, defol_status)
}
\arguments{
\item{year}{An n-length integer vector of observation years}

\item{series}{An n-length factor or character vector of series names}

\item{gsi}{An n-length numeric vector of growth suppression index, such as
calculated by \code{\link[=gsi]{gsi()}}}

\item{ngsi}{An n-length numeric vector of normalized gsi, such as calculated
by \code{\link[=gsi]{gsi()}}.}

\item{defol_status}{An n-length factor or character vector denoting the
defoliation event status of each year. This uses a controlled vocabulary,
see \code{dfoliatR:::make_defol_status} for possible values.}
}
\value{
a tree-level \code{defol} object
}
\description{
Constructor for S3 defol class
}
