% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmj_h}
\alias{dmj_h}
\title{Demi John Douglas-fir}
\format{An \code{rwl} object with 17 tree-level series, standardized in ARSTAN.
Dates range from 1620-1997.}
\usage{
dmj_h
}
\description{
Western spruce budworm host series
}
\keyword{datasets}
