% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmj_nh}
\alias{dmj_nh}
\title{Demi John area ponderosa pine}
\format{
An \code{rwl} object with 1 series, 1675-1997.
}
\usage{
dmj_nh
}
\description{
Non-host pair to Demi John Douglas-fir \code{dmj_h}
}
\keyword{datasets}
