% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmj_obr}
\alias{dmj_obr}
\title{Demi John outbreak object}
\format{An object of class \code{obr} (inherits from \code{data.frame}) with 323 rows and 9 columns.}
\usage{
dmj_obr
}
\description{
Produced by running `outbreak(dmj_defol)`
}
\details{
@format An `outbreak` object with 323 rows and 9 columns
}
\keyword{datasets}
