% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ef_defol}
\alias{ef_defol}
\title{East Fork defol object}
\format{
A \code{defol} object with 2500 rows and 5 columns
}
\usage{
ef_defol
}
\description{
Produced by running \code{defoliate_trees(ef_h, ef_nh, series_end_event = TRUE)}
}
\keyword{datasets}
