% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ef_h}
\alias{ef_h}
\title{East Fork Jemez River Douglas-fir}
\format{An \code{rwl} object with 15 tree-level series, standardized in ARSTAN.
Dates range from 1776-1987.}
\usage{
ef_h
}
\description{
Western spruce budworm host series
}
\keyword{datasets}
