% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ef_nh}
\alias{ef_nh}
\title{East Fork Jemez area ponderosa pine}
\format{
An \code{rwl} chronology object with 1 series, 1612-1987. Standardized in
ARSTAN.
}
\usage{
ef_nh
}
\description{
Non-host pair chronology for East Fork Douglas-fir \code{ef_h}
}
\keyword{datasets}
