% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efk_defol}
\alias{efk_defol}
\title{East Fork defol object}
\format{
A \code{defol} object with 5142 rows and 5 columns
}
\usage{
efk_defol
}
\description{
Produced by running \code{defoliate_trees(efk_h, efk_nh, series_end_event = TRUE)}
}
\keyword{datasets}
