% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efk_h}
\alias{efk_h}
\title{East Fork Jemez River Douglas-fir}
\format{
An \code{rwl} object with 37 tree-level series, standardized in ARSTAN.
Dates range from 1776-1987.
}
\usage{
efk_h
}
\description{
Western spruce budworm host series
}
\references{
Swetnam, T.W., Lynch, A.M., 1993. Multicentury, Regional-Scale
Patterns of Western Spruce Budworm Outbreaks. Ecological Monographs 63,
399–424. https://doi.org/10.2307/2937153
}
\keyword{datasets}
