% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efk_obr}
\alias{efk_obr}
\title{East Fork outbreak object}
\format{
An \code{outbreak} object with 221 rows and 9 columns
}
\usage{
efk_obr
}
\description{
Produced by running \code{outbreak(efk_defol)}
}
\keyword{datasets}
