% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{events_table}
\alias{events_table}
\title{Create a runs table of events}
\usage{
events_table(status, events)
}
\arguments{
\item{status}{vector of defoliation or outbreak status}

\item{events}{vector of events types to include in the table}
}
\description{
Create a runs table of events
}
