% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gsi}
\alias{gsi}
\title{Calculate the growth suppression index}
\usage{
gsi(input_series)
}
\arguments{
\item{input_series}{data.frame rwl object with the host tree series as the
first column and the non-host chronology as the second. Years should be
\code{row.names}.}
}
\value{
data.frame with the two input columns (host and nonhost series) and 3
  added columns:

  \itemize{
  \the mean/sd adjusted non-host chronology,
  \the growth suppression index (gsi) of the host series after subtraction of the
  adjusted nonhost chronology,
  \the normalized growth suppression index
  (ngsi) generated by \code{scale} on the gsi.
  }
}
\description{
This function removes the nonhost growth signal from a host series.
}
\details{
The growth suppression index (GSI) is referred to as the "corrected"
  series in OUTBREAK. It is calculated as:

  \deqn{GSI(i) = H(i) - (NH(i) - mean(NH)) sd(H)/sd(NH)}

  where H and NH are the host and nonhost tree-ring series as standardized
  index values; i is the year, and the functions \code{mean} and \code{sd}
  are applied to the common period.

  \code{gsi} will be rarely run directly by a user. It is called from
  \code{defoliate_trees}.
}
