% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.defol}
\alias{is.defol}
\title{Check if object is tree-level defoliation object: `defol`}
\usage{
is.defol(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Boolean indicating whether `x` is a defol object.
}
\description{
Check if object is tree-level defoliation object: `defol`
}
