% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.obr}
\alias{is.obr}
\title{Check if object is outbreak, meaning site-level outbreak object}
\usage{
is.obr(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Boolean indicating whether `x` is an outbreak object.
}
\description{
Check if object is outbreak, meaning site-level outbreak object
}
