% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_defol_status}
\alias{make_defol_status}
\title{Create factor for `defol_status`}
\usage{
make_defol_status(x)
}
\arguments{
\item{x}{a vector of defol types}
}
\value{
A factor with appropriate defol levels
}
\description{
Create factor for `defol_status`
}
