% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_defol}
\alias{new_defol}
\title{Constructor for S3 defol class}
\usage{
new_defol(year, series, gsi, ngsi, defol_status)
}
\arguments{
\item{year}{An integer vector of measured years for each \code{series}}

\item{series}{A factor of series names for each \code{year}}

\item{gsi}{A numeric vector of growth suppression index from \code{gsi()}}

\item{ngsi}{A numeric vector of normalized gsi from \code{gsi()}}
}
\value{
a tree-level `defol` object
}
\description{
Constructor for S3 defol class
}
