% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.defol}
\alias{plot.defol}
\title{Plot a \code{defol} object}
\usage{
\method{plot}{defol}(...)
}
\arguments{
\item{...}{arguments passed to \code{plot_defol}}
}
\description{
Plot a \code{defol} object
}
