% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_defol}
\alias{plot_defol}
\title{Graph individual tree-ring series from a 'defol' object}
\usage{
plot_defol(x, col_defol = "black")
}
\arguments{
\item{x}{a 'defol' object}

\item{col_defol}{the fill color of the verticle bars indicating defoliation years}
}
\description{
Graph individual tree-ring series from a 'defol' object
}
