% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sample_depth}
\alias{sample_depth}
\title{Calculate the sample depth of a defol object}
\usage{
sample_depth(x)
}
\arguments{
\item{x}{A \code{defol} object.}
}
\value{
A data.frame containing the years and number of trees
}
\description{
Calculate the sample depth of a defol object
}
\examples{
data("dmj_defol")
head(sample_depth(dmj_defol))

}
