% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stack_defoliation}
\alias{stack_defoliation}
\title{stack defolation list object}
\usage{
stack_defoliation(x)
}
\arguments{
\item{x}{a list object created within \code{defoliate_trees}.}
}
\value{
a data.frame defol object
}
\description{
stack defolation list object
}
