% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defoliate_trees.R
\name{defoliate_trees}
\alias{defoliate_trees}
\title{Measure defoliation events in host trees}
\usage{
defoliate_trees(host_tree, nonhost_chron, duration_years = 8,
  max_reduction = -1.28, list_output = FALSE)
}
\arguments{
\item{host_tree}{a data.frame rwl object containing the tree-level growth series for all
host trees to be compared to the non-host chronology}

\item{nonhost_chron}{a data.frame rwl object comtaining a single non-host chronology}

\item{duration_years}{the mimimum number of years in which to consider a defolation event}

\item{max_reduction}{the minimum level of tree growth to be considered in defoliation}

\item{list_output}{defaults to \code{FALSE}. This option is to output a long list object containing a separate data.frame for each series in
\code{host_tree} that includes the input series and the \code{nonhost_chron}, the corrected series, and
the character string identifying the defoliation events.}
}
\value{
a list object with elements containing data.frame rwl objects of the host and non-host series, corrected
}
\description{
Conduct host/nonhost comparison to identify defoliation events
}
