% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ef_obr}
\alias{ef_obr}
\title{East Fork outbreak object}
\format{An \code{outbreak} object with 221 rows and 9 columns}
\usage{
ef_obr
}
\description{
Produced by running \code{outbreak(ef_defol)}
}
\keyword{datasets}
