% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efk_h}
\alias{efk_h}
\title{East Fork Jemez River Douglas-fir}
\format{
An \code{rwl} object with 37 tree-level series, standardized in ARSTAN.
Dates range from 1776-1987.
}
\source{
\url{https://esajournals.onlinelibrary.wiley.com/doi/10.2307/2937153}
}
\usage{
efk_h
}
\description{
Western spruce budworm host series
}
\keyword{datasets}
