% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efk_nh}
\alias{efk_nh}
\title{Baca ponderosa pine chronology}
\format{
An \code{rwl} chronology object with 1 series, 1612-1987. Standardized in
ARSTAN.
}
\source{
\url{https://esajournals.onlinelibrary.wiley.com/doi/10.2307/2937153}
}
\usage{
efk_nh
}
\description{
Non-host pair chronology for East Fork Douglas-fir \code{efk_h}
}
\keyword{datasets}
