% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_defoliation}
\alias{id_defoliation}
\title{Identify defoliation events in a host series}
\usage{
id_defoliation(input_series, duration_years = 8, max_reduction = -1.28,
  end_series_event = FALSE)
}
\arguments{
\item{input_series}{a data.frame with 5 columns, as generated by \code(gsi).}

\item{duration_years}{the minimum length of time in which the tree is
considered to be in defoliation}

\item{max_reduction}{minimum values of ngsi allowed within a defoliation
event. If a sequence of negative ngsi values does not reach this threshold,
the potential event is rejected. Defaults to -1.28}

\item{end_series_event}{Binary, defaults to FALSE. Whether to consider 1-2
possitive index values at the end of the series as part of an ongoing
defoliation event}
}
\value{
after performing runs analyses, the function adds a column to the
  input data.frame that distinguishes years of defoliation and the maximum
  defoliation year (ie. the year the greatest negative growth departure
  within the event).
}
\description{
Identify defoliation events in a host series
}
