% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{outbreak}
\alias{outbreak}
\title{Composite defoliation series to determine outbreak events}
\usage{
outbreak(x, filter_perc = 25, filter_min_series = 3,
  filter_min_defol = 1, force = FALSE)
}
\arguments{
\item{x}{a defol object}

\item{filter_perc}{the minimum percentage of defoliated trees to be
considered an outbreak. Default is 25 percent.}

\item{filter_min_series}{The minimum number of trees required for an outbreak
event. Default is 3 trees.}

\item{filter_min_defol}{The minimum number of trees recording a defoliation
event. Default is 1 tree.}

\item{force}{Ignores the class structure if the data frame, to aid in  batch
processing. Defaults to FALSE. Use cautiously.}
}
\description{
Composite defoliation series to determine outbreak events
}
