% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{outbreak_stats}
\alias{outbreak_stats}
\title{Outbreak statistics}
\usage{
outbreak_stats(x)
}
\arguments{
\item{x}{An outbreak object after running \code{outbreak}}
}
\value{
A data.frame with descriptive statistics for each outbreak event determined by \code{outbreak},
 inluding start and end years, duration, the year with the most number of trees in the outbreak and its
 associated tree count, and the year with the maximum growth suppression with its associated mean_index value.
}
\description{
Outbreak statistics
}
