% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_defol}
\alias{plot_defol}
\title{Produce a Gantt plot of individual tree-ring series to show defoliation
events in time}
\usage{
plot_defol(x, breaks)
}
\arguments{
\item{x}{a \code{defol} object produced by \code{\link[=defoliate_trees]{defoliate_trees()}}.}

\item{breaks}{a vector length two providing threshold (negative) \code{ngsi}
values to separate minor, moderate, and severe defoliation events. If
blank, the mean and 1st quartile are used.}
}
\description{
Produce a Gantt plot of individual tree-ring series to show defoliation
events in time
}
\examples{
data("dmj_defol")
plot_defol(dmj_defol)

## Change the values severity classes
plot_defol(dmj_defol, breaks = c(-1.0, -0.5))

}
