% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_outbreak}
\alias{plot_outbreak}
\title{Produce a stacked plot to present composited, site-level insect outbreak
chronologies}
\usage{
plot_outbreak(x, disp_index = "mean_ngsi")
}
\arguments{
\item{x}{an 'outbreak' object produced by \code{outbreak}}

\item{disp_index}{Identify the timeseries index to plot. Defaults to
\code{mean_ngsi}, the average normalized growth suppression index for the
site. The only other option is \code{mean_gsi}, the average growth suppression index.}
}
\description{
Produce a stacked plot to present composited, site-level insect outbreak
chronologies
}
