% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_outbreak}
\alias{plot_outbreak}
\title{Produce a stacked plot to present composited, site-level insect outbreak chronologies}
\usage{
plot_outbreak(x)
}
\arguments{
\item{x}{an 'outbreak' object produced by \code{outbreak}}
}
\description{
Produce a stacked plot to present composited, site-level insect outbreak chronologies
}
