% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stack_defoliation}
\alias{stack_defoliation}
\title{Stack a defoliation list}
\usage{
stack_defoliation(x)
}
\arguments{
\item{x}{a list object created by \code{\link[=defoliate_trees]{defoliate_trees()}}}
}
\value{
a defol object (long-format data frame)
}
\description{
Stack a defoliation list
}
